+++ packages/codenic_exception_converter/CHANGELOG.md
@@ -1,6 +1,12 @@
+## 1.0.2
+- Bump dependencies
+- Use `FutureOr` in `ExceptionConverter.observe` and 
+`ExceptionConverter.observe` `task` parameters.
+- Deprecate `ExceptionConverter.observeSync` and 
+`ExceptionConverterSuite.observeSync`
+
 ## 1.0.1
 - Update documentation link in README.md
 
 ## 1.0.0
-
 - Initial version